% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoolNet_CellCycle_Network.R
\docType{data}
\name{BoolNet_CellCycle_Network}
\alias{BoolNet_CellCycle_Network}
\title{BoolNet CellCycle Network}
\format{
A traditional Boolean type of cell cycle network
}
\usage{
data(BoolNet_CellCycle_Network)
}
\description{
Data generated via BoolNet::BoolNet::loadNetwork
}
\examples{
data(BoolNet_CellCycle_Network)
BoolNet_CellCycle_Network
}
\references{
Mussel, Hopfensitz et al. 2010, BoolNet - an R package
 for generation, reconstruction and analysis of Boolean networks
}
\keyword{cell}
\keyword{cycle}
\keyword{network}
