% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildProbabilityTreeOnTargetGene}
\alias{buildProbabilityTreeOnTargetGene}
\title{build Probability Tree On the targetGene}
\usage{
buildProbabilityTreeOnTargetGene(
  targetGene,
  mainParameters,
  genes,
  matchedgenes,
  matchedexpression,
  maxK = 4L,
  temporal = 1L,
  targetCounts = NULL,
  findPositiveRegulate = FALSE,
  findNegativeRegulate = FALSE
)
}
\arguments{
\item{targetGene}{The target gene}

\item{mainParameters}{An environment variable holds all input data}

\item{genes}{All conditional genes}

\item{matchedgenes}{processed genes}

\item{matchedexpression}{matched the expression of an interaction}

\item{maxK}{The maximum level that can be drilled in.}

\item{temporal}{The temporal time steps}

\item{targetCounts}{Counts the number of target gene in the data}

\item{findPositiveRegulate}{Optional if TRUE find positive regulation only}

\item{findNegativeRegulate}{Optional if TRUE find negative regulation only}
}
\description{
build Probability Tree On the targetGene
}
