% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_FBN.R
\name{checkNumeric}
\alias{checkNumeric}
\title{A simple function to check a value is numeric or not}
\usage{
checkNumeric(x)
}
\arguments{
\item{x}{A value that need to be checked.}
}
\value{
Error or NULL
}
\description{
A simple function to check a value is numeric or not
}
