% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_FBN.R
\name{checkProbabilityTypeData}
\alias{checkProbabilityTypeData}
\title{A simple function to check a value is probability type
data.}
\usage{
checkProbabilityTypeData(x)
}
\arguments{
\item{x}{A value that need to be checked.}
}
\value{
Error or NULL
}
\description{
A simple function to check a value is probability type
data.
}
