% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_FBN.R
\name{CheckRightTypeTimeseriesData}
\alias{CheckRightTypeTimeseriesData}
\title{This function is used to check whether or not the data 
is the right type for \code{FBNNet}.}
\usage{
CheckRightTypeTimeseriesData(timeseries_data)
}
\arguments{
\item{timeseries_data}{The timeseries data}
}
\value{
Error or NULL
}
\description{
This function is used to check whether or not the data 
is the right type for \code{FBNNet}.
}
