% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_FBN.R
\name{checkSimilarity}
\alias{checkSimilarity}
\title{Check the similarity between time series}
\usage{
checkSimilarity(originalTimeseriesCube, reconstructedTimeSeriesCube)
}
\arguments{
\item{originalTimeseriesCube}{The original data set that
contains samples and each sample contains genes and time points}

\item{reconstructedTimeSeriesCube}{The reconstructed data set
that contains samples and each sample contains genes and time points}
}
\value{
similarity report
}
\description{
Check the similarity between time series
}
\examples{
##coming later
}
