% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_FBN.R
\name{clusterdDiscreteData}
\alias{clusterdDiscreteData}
\title{A function to cluster data}
\usage{
clusterdDiscreteData(
  originalTimeSeriesData,
  discretedTimeSeriesdata,
  minElements = 10,
  maxElements = 30,
  membexp = 2
)
}
\arguments{
\item{originalTimeSeriesData}{The original expression timeseries data}

\item{discretedTimeSeriesdata}{The discreted timeseries data}

\item{minElements}{The minimum elements of a cluster}

\item{maxElements}{The maximum elements of a cluster}

\item{membexp}{number r strictly larger than 1 specifying the
membership exponent used in the fit criterion; see the ‘Details’
below. Default: 2 which used to be hardwired inside FANNY.
@return A object of ClusteredTimeseriesData}
}
\description{
A function to cluster data
}
