% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_FBN.R
\name{constructFBNCube}
\alias{constructFBNCube}
\title{Create an Orchard cube}
\usage{
constructFBNCube(
  target_genes,
  conditional_genes,
  timeseriesCube,
  maxK = 5,
  temporal = 1,
  useParallel = FALSE
)
}
\arguments{
\item{target_genes}{A vector of genes that will be treated as target genes}

\item{conditional_genes}{All genes that are available for building up the
cube}

\item{timeseriesCube}{A list of samples in which a sample is a matrix that
contains gene states 
where genes in rows and time points in columms}

\item{maxK}{The maximum level the cube can dig in}

\item{temporal}{A value that used to be 1 indicates the previous steps the
current one can depend on}

\item{useParallel}{If it is TRUE, the constructing will run it in parallel,
otherwise in a singl thread}
}
\value{
A Orcahrd cube that contains all precomputed measures
}
\description{
This is the main function(s) to genereate a sigle Orchard Cube or a group
of cubes
}
\examples{
require(BoolNet)
data('ExampleNetwork')
initialStates <- generateAllCombinationBinary(ExampleNetwork$genes)
trainingseries <- genereateBoolNetTimeseries(ExampleNetwork,
                                           initialStates,
                                           43,
                                           type='synchronous')
cube<-constructFBNCube(target_genes = ExampleNetwork$genes,
                       conditional_genes = ExampleNetwork$genes,
                       timeseriesCube = trainingseries,
                       maxK = 4,
                       temporal = 1,
                       useParallel = FALSE)
network <- mineFBNNetwork(cube,ExampleNetwork$genes)
network
## draw the general graph
FBNNetwork.Graph(network)
}
