% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_FBN.R
\name{constructFBNCubeAndNetworkInClusters}
\alias{constructFBNCubeAndNetworkInClusters}
\title{Create an Orchard cube with clusters}
\usage{
constructFBNCubeAndNetworkInClusters(
  clusteredTimeseriesData,
  maxK = 5,
  temporal = 1,
  useParallel = FALSE,
  threshold_confidence = 1,
  threshold_error = 0,
  threshold_support = 1e-05,
  maxFBNRules = 20,
  output_cube = FALSE,
  parallel_on_group = FALSE
)
}
\arguments{
\item{clusteredTimeseriesData}{clustered timeseries data}

\item{maxK}{The maximum level the cube can dig in.}

\item{temporal}{A value that used to be 1 indicates the previous steps the
current one can depend on.}

\item{useParallel}{If it is TRUE, the constructing will run it in parallel,
otherwise in a singl thread.}

\item{threshold_confidence}{A threshold of confidence.}

\item{threshold_error}{A threshold of error.}

\item{threshold_support}{A threshold of support.}

\item{maxFBNRules}{A threshold of maximum rules}

\item{output_cube}{Optional, if it is TRUE, the cube and networks will be
returned, otherwise, only networks are returned.}

\item{parallel_on_group}{Optional, if it is TRUE, then the parallelisation
will be on cluster groups.}
}
\value{
A Orcahrd cube that contains all precomputed measures
}
\description{
This is the main function(s) to genereate a sigle Orchard Cube or a group
of cubes.
}
