% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utility_FBN.R
\name{constructFBNFunctions}
\alias{constructFBNFunctions}
\title{An internal function to construct the FBN functions
with an expression tree.}
\usage{
constructFBNFunctions(expressionTree)
}
\arguments{
\item{expressionTree}{A expression tree}
}
\description{
An internal function to construct the FBN functions
with an expression tree.
}
