% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_FBN.R
\name{convert_to_NGO}
\alias{convert_to_NGO}
\title{Convert To Dynamic Network Graphic Object}
\usage{
convert_to_NGO(timeseries, FBNnetwork, networkobject)
}
\arguments{
\item{timeseries}{The timeseries data}

\item{FBNnetwork}{The fundamental Boolean network.}

\item{networkobject}{The fundamental Boolean network
objects.}
}
\value{
A dataframe object.
}
\description{
internal method to convert the timeseries data 
into the graphic object. 
(ConvertToDynamicNetworkGraphicObject)
}
