% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{convertInteraction}
\alias{convertInteraction}
\title{A function to convert the traditional network interactions
to fundamental Boolean network interactions.}
\usage{
convertInteraction(interaction)
}
\arguments{
\item{interaction}{the traditional network interactions}
}
\description{
A function to convert the traditional network interactions
to fundamental Boolean network interactions.
}
