% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utility_FBN.R
\name{convertIntoExpressionTree}
\alias{convertIntoExpressionTree}
\title{An internal function to convert a splited Boolean expression
into an expression tree.}
\usage{
convertIntoExpressionTree(splitedExpression)
}
\arguments{
\item{splitedExpression}{A splited Boolean expression}
}
\description{
An internal function to convert a splited Boolean expression
into an expression tree.
}
