% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{convertIntoSampleTimeSeries}
\alias{convertIntoSampleTimeSeries}
\title{This method is used to converts normalized timeseries data into a list of samples}
\usage{
convertIntoSampleTimeSeries(
  normalizedData,
  func = function(x) paste(x[1], x[2], x[3], sep = "-"),
  splitor = as.character("-")
)
}
\arguments{
\item{normalizedData}{An output of the method normalizeTimesereisRawData that normalized timeseries data}

\item{func}{A function that specified how to split the column names}
}
\description{
This method is used to converts normalized timeseries data into a list of samples
}
