% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utility_FBN.R
\name{convertMinedResultToFBNNetwork}
\alias{convertMinedResultToFBNNetwork}
\title{convert the mined result to FBNNetwork object}
\usage{
convertMinedResultToFBNNetwork(minerresult, genes)
}
\arguments{
\item{minerresult}{The mined result from FBNmine.R}

\item{genes}{The target genes in the output}
}
\value{
A object of FBN network
}
\description{
convert the mined result to FBNNetwork object
}
\examples{
mat1<-matrix(c('1','0','0','1','0','0','0','1','1'),3,3, dimnames=list(c('gene1','gene2','gene3'),c('1','2','3')))
mat2<-matrix(c('1','1','0','1','0','1','1','1','0'),3,3, dimnames=list(c('gene1','gene2','gene3'),c('1','2','3')))
listtest<-list(mat1,mat2)
cube<-constructFBNCube(c('gene1','gene2'),c('gene1','gene2','gene3'),listtest,4,1,FALSE)
res<-mineFBNNetworkCore(cube,c('gene1','gene2'))
res<-mineFBNNetworkStage2(res)
finalresult<-outputMinerResultToFBNNetwork(res,c('gene1','gene2'))
}
