% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNBioinformatics.R
\name{convertTimeseriesProbsetNameToGeneName}
\alias{convertTimeseriesProbsetNameToGeneName}
\title{A method that map the row names (probesets) of a specific timeseries data into gene names
Old function name mapPropersetsWithhgu133plus2Db}
\usage{
convertTimeseriesProbsetNameToGeneName(orderSampleTimeSeries)
}
\arguments{
\item{orderSampleTimeSeries}{A sorted time series data, which is the output of the method reorderSampleTimeSeries}
}
\description{
A method that map the row names (probesets) of a specific timeseries data into gene names
Old function name mapPropersetsWithhgu133plus2Db
}
