% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{convertToBooleanNetworkCollection}
\alias{convertToBooleanNetworkCollection}
\title{A function to convert the traditional Boolean network to
fundamental Boolean network collection.}
\usage{
convertToBooleanNetworkCollection(network)
}
\arguments{
\item{network}{The traditional Boolean network.}
}
\description{
A function to convert the traditional Boolean network to
fundamental Boolean network collection.
}
