% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_FBN.R
\name{ConvertToDynamicNetworkGraphicObject}
\alias{ConvertToDynamicNetworkGraphicObject}
\title{internal method to convert the timeseries data 
into the graphic object.}
\usage{
ConvertToDynamicNetworkGraphicObject(timeseries, FBNnetwork, networkobject)
}
\arguments{
\item{timeseries}{The timeseries data}

\item{FBNnetwork}{The fundamental Boolean network.}

\item{networkobject}{The fundamental Boolean network
objects.}
}
\value{
A dataframe object.
}
\description{
internal method to convert the timeseries data 
into the graphic object.
}
