% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utility_FBN.R
\name{convertToFBNNetwork}
\alias{convertToFBNNetwork}
\title{Convert a boolean network object to fundamental boolean function object}
\usage{
convertToFBNNetwork(network)
}
\arguments{
\item{network}{A boolean network object}
}
\value{
An object of FBN network
}
\description{
Convert a boolean network object to fundamental boolean function object
}
\examples{
coming later
}
