% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_data_FBN.R
\name{discreteTimeSeries}
\alias{discreteTimeSeries}
\title{A method to discrete timeseries data based on the threshod of mean value}
\usage{
discreteTimeSeries(timeSeriesData, method = c("average", "distance"))
}
\arguments{
\item{timeSeriesData}{A list of time series data containing original value}

\item{method}{A method to discrete the timeseries data.}
}
\description{
A method to discrete timeseries data based on the threshod of mean value
}
