% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_FBN.R
\name{dissolve}
\alias{dissolve}
\title{A function moves the sub list items to its parent list}
\usage{
dissolve(x)
}
\arguments{
\item{x}{A complex list}
}
\value{
A simplified list
}
\description{
This function is used to simplify the complex list
}
