% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{dividedDataIntoSmallgroups}
\alias{dividedDataIntoSmallgroups}
\title{An internal function that divides large data into small groups.}
\usage{
dividedDataIntoSmallgroups(discretedTimeSeriesdata, maxElements = 20, maxK = 4)
}
\arguments{
\item{discretedTimeSeriesdata}{discreted timeseries data}

\item{maxElements}{the max elements to divide the discreted timeseries data into}

\item{maxK}{The max levels.}
}
\value{
A list of data
}
\description{
An internal function that divides large data into small groups.
}
