% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{dividedDataIntoSubgroups}
\alias{dividedDataIntoSubgroups}
\title{An function to divide large data into small groups}
\usage{
dividedDataIntoSubgroups(discretedTimeSeriesdata, maxElements = 20, maxK = 4)
}
\arguments{
\item{discretedTimeSeriesdata}{discreted timeseries data}

\item{maxElements}{the max elements to divide the 
discreted timeseries data into}

\item{maxK}{The maximum level that can be drilled into}
}
\description{
An function to divide large data into small groups
}
