% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_FBN.R
\name{dividedintosmallgroups}
\alias{dividedintosmallgroups}
\title{An internal function to divide large data in cluster groups}
\usage{
dividedintosmallgroups(
  timeSeriesData,
  minElements = 10,
  maxElements = 30,
  membexp = 2
)
}
\arguments{
\item{timeSeriesData}{The original expression timeseries data}

\item{minElements}{The minimum elements of a cluster}

\item{maxElements}{The maximum elements of a cluster}

\item{membexp}{number r strictly larger than 1 specifying the
membership exponent used in the fit criterion; see the ‘Details’
below. Default: 2 which used to be hardwired inside FANNY.
@return The result of \code{fuzzyTreeCluster}}
}
\description{
An internal function to divide large data in cluster groups
}
