% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{dividedVectorIntoSmallgroups}
\alias{dividedVectorIntoSmallgroups}
\title{An internal function that divides large data into small groups.}
\usage{
dividedVectorIntoSmallgroups(vector, maxElements = 20)
}
\arguments{
\item{vector}{A vector that need to be splited}

\item{maxElements}{the max elements to divide the discreted timeseries data into}
}
\value{
A list of data
}
\description{
An internal function that divides large data into small groups.
}
