% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_FBN.R
\name{doNonParallelWork}
\alias{doNonParallelWork}
\title{Executing a list of processes not in parallel}
\usage{
doNonParallelWork(parallelFuc, listitems, ...)
}
\arguments{
\item{parallelFuc}{A process that will be run in parallel}

\item{listitems}{The main list for each process}

\item{...}{other parameters}
}
\value{
No return
}
\description{
Executing a list of processes not in parallel
}
\examples{
coming later
}
