% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_FBN.R
\name{doNonParallelWorkDecrease}
\alias{doNonParallelWorkDecrease}
\title{Executing a list of processes in parallel with decreasing list}
\usage{
doNonParallelWorkDecrease(parallelFuc, listitems, unprocessedListitems, ...)
}
\arguments{
\item{parallelFuc}{A process that will be run in parallel}

\item{listitems}{The main list for each process}

\item{unprocessedListitems}{The remain list items that haven't been processed yet}

\item{...}{other parameters}
}
\value{
No return
}
\description{
Executing a list of processes in parallel with decreasing list
}
\examples{
coming later
}
