% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleTimeseriesData.R
\docType{data}
\name{ExampleTimeseriesData}
\alias{ExampleTimeseriesData}
\title{Fundamental Boolean model Example}
\format{
An object of class \code{'list'};
}
\usage{
data(ExampleTimeseriesData)
}
\description{
Example described in the paper 'A Novel Data-Driven Boolean Model for Genetic Regulatory Networks' (Chen, L., etc., 2018)
}
\examples{
data(ExampleTimeseriesData)
ExampleTimeseriesData
}
\references{
Chen et al.(2018), Front. Physiol., 25 September 2018, 
(\href{https://doi.org/10.3389/fphys.2018.01328}{Front. Physiol.})
}
\keyword{Boolean}
\keyword{Fundamental}
\keyword{Modelling}
