% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extractGeneStateFromTimeSeriesCube}
\alias{extractGeneStateFromTimeSeriesCube}
\title{A function that extract gene states from time series cube}
\usage{
extractGeneStateFromTimeSeriesCube(timeSeriesCube, temporal)
}
\arguments{
\item{timeSeriesCube}{The time series cube}

\item{temporal}{The temporal time step}
}
\description{
A function that extract gene states from time series cube
}
