% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNBioinformatics.R
\name{FBNBioinformatics}
\alias{FBNBioinformatics}
\title{Functions related to Bioinformatics}
\arguments{
\item{files}{The file folder path or directory that
contains the affy files (*.cel)}

\item{useGCRMA}{optional, if true it use GCRMA method to normalize
the affy data, otherwise, use the default RMA method.
 GCRMA, A bias-corrected RMA; FALSE, use RMA, which is based on
  Robust Multi-Chip' average}

\item{cdfname}{The name of CDF that is associated with the affy data.}

\item{cellDirectory}{A directory that contains affy raw data}

\item{outPutEvalue}{optional to write the result into disk}

\item{rawData}{the raw data that is the output from the method
getRelatedAffyRawData}

\item{ges_assess_no}{the GSE identity id, such as GSE35635}
}
\description{
A collection of functions that used to handle
bioinformatics related tasks
}
\examples{
# exprs <- downloadGenomeDataExprs(GSE35635)
# exprs
}
\references{
Chen et al.(2018), Front. Physiol., 25 September 2018,
(\href{https://doi.org/10.3389/fphys.2018.01328}{Front. Physiol.})

Bioconductor (2019), https://www.bioconductor.org/
}
\author{
Leshi Chen, leshi, chen@lincolnuni.ac.nz, chenleshi@hotmail.com
}
\keyword{Boolean}
\keyword{Fundamental}
\keyword{Genetic}
\keyword{Network}
\keyword{Network,}
\keyword{Regulatory}
