% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_FBN.R
\name{FBNDataReduction}
\alias{FBNDataReduction}
\title{Remove duplicate timeseries sample data}
\usage{
FBNDataReduction(timeseriescube)
}
\arguments{
\item{timeseriescube}{A list of time series data}
}
\value{
The return a list of time series data
}
\description{
Remove duplicate timeseries sample data
}
\examples{
mat1<-matrix(c('1','2','3','4','5','6','7','8','9'),3,3)
mat2<-matrix(c('1','2','3','4','5','6','7','8','9'),3,3)
listtest<-list(mat1,mat2)
FBNDataReduction(listtest)
}
