% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNExampleNetworks.R
\docType{data}
\name{FBNExampleNetworks}
\alias{FBNExampleNetworks}
\title{Example Network data}
\format{
A simple network
}
\usage{
data(FBNExampleNetworks)
}
\description{
Data generated for testing purpose.
}
\examples{
data(FBNExampleNetworks)
FBNExampleNetworks
}
\references{
Chen, L., D. Kulasiri, and S. Samarasinghe,
 A Novel Data-Driven Boolean Model for Genetic Regulatory
  Networks. Front Physiol, 2018. 9: p. 1328
(\href{https://www.frontiersin.org/articles/10.3389/fphys.2018.01328/full/}{frontiersin})
}
\keyword{Example}
\keyword{Network}
