% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_FBN.R
\name{FBNNetwork.Graph.ConvertToNetworkGraphicObject}
\alias{FBNNetwork.Graph.ConvertToNetworkGraphicObject}
\title{This method is used to convert the FBN network into a FBN networkobject for graphs}
\usage{
FBNNetwork.Graph.ConvertToNetworkGraphicObject(
  FBNnetwork,
  david_gene_list = NULL,
  show_decay = FALSE
)
}
\arguments{
\item{FBNnetwork}{A FBN network object}

\item{david_gene_list}{A data frame of DAVID biological annotation}

\item{show_decay}{Optional, if TRUE, will draw the connections for decay.}
}
\value{
An network graphic object
}
\description{
This method is used to convert the FBN network into a FBN networkobject for graphs
}
\examples{
##coming later
}
