% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_FBN.R
\name{FBNNetwork.Graph.DrawAttractor}
\alias{FBNNetwork.Graph.DrawAttractor}
\title{A method to draw attractors}
\usage{
FBNNetwork.Graph.DrawAttractor(
  fbnNetwork,
  FBMAttractors,
  index = 1,
  networkobject = NULL,
  dynamicnetworkobject = NULL
)
}
\description{
A method to draw attractors
}
