% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_FBN.R
\name{FBNNetwork.Graph.DrawingAttractorInternal}
\alias{FBNNetwork.Graph.DrawingAttractorInternal}
\title{An Internal method to draw atractor internally}
\usage{
FBNNetwork.Graph.DrawingAttractorInternal(
  networkobject,
  dynamicNetworkGraphicObject,
  matrix
)
}
\arguments{
\item{networkobject}{Fundamental Boolean Network objects.}

\item{dynamicNetworkGraphicObject}{dynamicNetworkGraphicObject.}

\item{matrix}{The timeseries matrix}
}
\description{
An Internal method to draw atractor internally
}
