% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_FBN.R
\name{FBNNetwork.Graph.StaticNetwork}
\alias{FBNNetwork.Graph.StaticNetwork}
\title{Display a static network in a slice}
\usage{
FBNNetwork.Graph.StaticNetwork(networkobject, david_gene_list = NULL)
}
\arguments{
\item{david_gene_list}{the list object of DAVID annotation gene list}

\item{FBNnetwork}{A FBN network object}
}
\value{
No return
}
\description{
Display a static network in a slice
}
\examples{
coming later
}
