% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_FBN.R
\name{FBNNetwork.Graph.StaticNetworkInSlice}
\alias{FBNNetwork.Graph.StaticNetworkInSlice}
\title{Display a dynamic network in a slice}
\usage{
FBNNetwork.Graph.StaticNetworkInSlice(
  networkobject,
  timepoint,
  dynamicNetworkGraphicObject
)
}
\arguments{
\item{networkobject}{A net work object}

\item{timepoint}{The target time point}

\item{dynamicNetworkGraphicObject}{A dynamic network graphic object}
}
\value{
No return
}
\description{
Display a dynamic network in a slice
}
\examples{
coming later
}
