% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{filterNetworkConnections}
\alias{filterNetworkConnections}
\title{filtering out non regulate genes}
\usage{
filterNetworkConnections(networks)
}
\arguments{
\item{networks}{The FBN networks that are going to be filtered}
}
\value{
filtered networks
}
\description{
filtering out non regulate genes
}
