% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{filterNetworkConnectionsByGenes}
\alias{filterNetworkConnectionsByGenes}
\title{filtering networks with specific genes}
\usage{
filterNetworkConnectionsByGenes(
  networks,
  genelist = c(),
  exclusive = TRUE,
  expand = FALSE
)
}
\arguments{
\item{networks}{The FBN networks that are going to be filtered}
}
\value{
filtered networks
}
\description{
filtering networks with specific genes
}
