% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{filterNetworkConnectionsByInputGenes}
\alias{filterNetworkConnectionsByInputGenes}
\title{Find Backward Network By Genes}
\usage{
filterNetworkConnectionsByInputGenes(networks, genelist = c())
}
\arguments{
\item{networks}{The Fundamental Boolean Network internally}

\item{genelist}{A list of genes}

\item{exclusive}{Optional}

\item{expand}{Optional}
}
\description{
Find Backward Network By Genes
}
