% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{findAllBackwardRelatedGenes}
\alias{findAllBackwardRelatedGenes}
\title{Find Backward Network By Genes}
\usage{
findAllBackwardRelatedGenes(
  networks,
  target_gene,
  regulationType = NULL,
  target_type = NULL,
  maxDeep = 1,
  next_level_mix_type = FALSE
)
}
\arguments{
\item{networks}{The Fundamental Boolean Network}

\item{target_gene}{The target gene}

\item{regulationType}{The type of regulation either in 1 (Up regulation) or 0 (down regulation)}

\item{target_type}{The target connection type}

\item{maxDeep}{The maximum layous that need to be drilled down.}

\item{next_level_mix_type}{If TRUE, the next level will contain all type of connections.}
}
\description{
Find Backward Network By Genes
}
