% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{findAllForwardRelatedGenes}
\alias{findAllForwardRelatedGenes}
\title{Find Forward related Network By Genes internal}
\usage{
findAllForwardRelatedGenes(
  networks,
  target_gene,
  regulationType = NULL,
  target_type = NULL,
  main_target_gene = NULL,
  main_target_type = NULL,
  maxDeep = 1,
  next_level_mix_type = next_level_mix_type
)
}
\arguments{
\item{networks}{The Fundamental Boolean Network}

\item{target_gene}{The target gene}

\item{regulationType}{The type of regulation either in 1 (Up regulation) or 0 (down regulation)}

\item{target_type}{The boolean type of target genes either in 1 (Activation) or 0 (Inhibition)}

\item{main_target_gene}{The main target gene.}

\item{main_target_type}{The main connection type.}

\item{maxDeep}{The maximum layous that need to be drilled down.}

\item{next_level_mix_type}{If TRUE, the next level will contain all type of connections.}
}
\description{
Find Forward related Network By Genes internal
}
