% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{findAllInputGenes}
\alias{findAllInputGenes}
\title{An internal function to find all input genes from a group of genes}
\usage{
findAllInputGenes(networkinteractions, genes)
}
\arguments{
\item{networkinteractions}{All network interactions}

\item{genes}{The target group genes.}
}
\description{
An internal function to find all input genes from a group of genes
}
