% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{findAllTargetGenes}
\alias{findAllTargetGenes}
\title{An internal function to find all target genes from a group of 
etwork interactions}
\usage{
findAllTargetGenes(networkinteractions)
}
\arguments{
\item{networkinteractions}{All network interactions}
}
\description{
An internal function to find all target genes from a group of 
etwork interactions
}
