% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{findBackwardRelatedNetworkByGenes}
\alias{findBackwardRelatedNetworkByGenes}
\title{Find Backward Network By Genes}
\usage{
findBackwardRelatedNetworkByGenes(
  networks,
  target_gene_list = c(),
  regulationType = NULL,
  maxDeep = 1,
  next_level_mix_type = FALSE
)
}
\arguments{
\item{networks}{The Fundamental Boolean Network}

\item{genelist}{The target genes}
}
\description{
Find Backward Network By Genes
}
