% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{findForwardRelatedNetworkByGenes}
\alias{findForwardRelatedNetworkByGenes}
\title{Find Forward related Network By Genes}
\usage{
findForwardRelatedNetworkByGenes(
  networks,
  target_gene_list = c(),
  regulationType = NULL,
  target_type = NULL,
  maxDeep = 1,
  next_level_mix_type = FALSE
)
}
\arguments{
\item{networks}{The Fundamental Boolean Network}

\item{target_gene_list}{The target genes}

\item{regulationType}{The type of regulation either in 1 (Up regulation) or 0 (down regulation)}

\item{target_type}{The boolean type of target genes either in 1 (Activation) or 0 (Inhibition)}
}
\description{
Find Forward related Network By Genes
}
