% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utility_FBN.R
\name{flatDeMorganLaw}
\alias{flatDeMorganLaw}
\title{An internal function to flat a splited Boolean expression
with DeMorganLaw}
\usage{
flatDeMorganLaw(splitedExpression)
}
\arguments{
\item{splitedExpression}{A splited Boolean expression}
}
\description{
An internal function to flat a splited Boolean expression
with DeMorganLaw
}
