% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_FBN.R
\name{fuzzyTreeCluster}
\alias{fuzzyTreeCluster}
\title{An internal function to do the fuzzy tree cluster}
\usage{
fuzzyTreeCluster(requireddata, minElements, maxElements, membexp = 3)
}
\arguments{
\item{requireddata}{The original expression timeseries data}

\item{minElements}{The minimum elements of a cluster}

\item{maxElements}{The maximum elements of a cluster}

\item{membexp}{number r strictly larger than 1 specifying the
membership exponent used in the fit criterion; see the ‘Details’
below. Default: 2 which used to be hardwired inside FANNY.
@return cluster data}
}
\description{
An internal function to do the fuzzy tree cluster
}
