% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{generateAllCombinationBinary}
\alias{generateAllCombinationBinary}
\title{generate all combaination binary data based on an vector of genes}
\usage{
generateAllCombinationBinary(genelist = c(), begin = 1, last = 0)
}
\arguments{
\item{genelist}{An vector of genes}

\item{begin}{The begin index}

\item{last}{The last index, the default is 0 means 2^length(genelist)}
}
\description{
generate all combaination binary data based on an vector of genes
}
\examples{
individualgenes<-c('a','b','c')
testdata2 <- generateAllCombinationBinary(individualgenes)
testdata2
}
