% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_FBN.R
\name{GenerateDynamicNetworkGraphicObject}
\alias{GenerateDynamicNetworkGraphicObject}
\title{An Internal method to generate dynamic network graphic objects}
\usage{
GenerateDynamicNetworkGraphicObject(
  networkobject,
  fromtimepoint,
  totimepoint,
  dynamicNetworkGraphicObject,
  timeseries
)
}
\arguments{
\item{networkobject}{Fundamental Boolean Network objects.}

\item{fromtimepoint}{A start point of the timeseries matrix.}

\item{totimepoint}{The end point of the timeseries matrix.}

\item{dynamicNetworkGraphicObject}{dynamicNetworkGraphicObject.}

\item{timeseries}{The timeseries matrix}
}
\description{
An Internal method to generate dynamic network graphic objects
}
