% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_FBN.R
\name{generateFBMNetwork}
\alias{generateFBMNetwork}
\title{Generate Fundamental Boolean Model type of Network}
\usage{
generateFBMNetwork(
  timeseries_data,
  method = c("kmeans", "edgeDetector", "scanStatistic"),
  maxK = 4,
  useParallel = FALSE,
  max_deep_temporal = 1,
  threshold_confidence = 1,
  threshold_error = 0,
  threshold_support = 1e-05,
  maxFBNRules = 5,
  network_only = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{timeseries_data}{A list of timeseries data of samples.}

\item{method}{Specify a method to discrete the data in the range of
('kmeans', 'edgeDetector', 'scanStatistic') for the function 
\code{BoolNet::binarizeTimeSeries} to convert the data from numeric value
 to boolean value.}

\item{maxK}{The maximum deep of the Orchard Cube can mine into.}

\item{useParallel}{optional, by default it is TRUE to run the network
inference algorithm in parallel. FALSE without parallel}

\item{max_deep_temporal, }{a setting for Temporal Fundamental Boolean model
that specifies the maximum temporal space}

\item{threshold_confidence}{A threshold of confidence (between 0 and 1) 
that used to filter the Fundamental Boolean functions}

\item{threshold_error}{A threshold of error rate (between 0 and 1) that used
to filter the Fundamental Boolean functions}

\item{threshold_support}{A threshold of support (between 0 and 1) that used
to filter the Fundamental Boolean functions}

\item{maxFBNRules}{The maximum rules per type (Activation and Inhibition)
per gene can be mined, the rest will be discarded}

\item{network_only}{optional for Debug purpose, if TRUE, only output the
 networks only, otherwise, output the Orchard cube as well. Warning, 
turn off this may cause memory leaking if the number of nodes is too large.
 divide them into subgroups.}

\item{verbose}{Optional, if it is TRUE, then output the logger information to
the console.}
}
\value{
An object of a list contains Fundamental Boolean Network, Orchard
 cube (optional) and discreted timeseries data
}
\description{
This is the main entry of the package FBNNet that can be used to mine the
 gene regulatory network.
}
\examples{
 
data('yeastTimeSeries')

network <- generateFBMNetwork(yeastTimeSeries, verbose = TRUE)
network
## draw the general graph
FBNNetwork.Graph(network)

## get the Orchard cube as well as networks
res <- generateFBMNetwork(yeastTimeSeries, network_only = FALSE)
res
}
\references{
Chen et al.(2018), Front. Physiol., 25 September 2018, 
(\href{https://doi.org/10.3389/fphys.2018.01328}{Front. Physiol.})

Mussel, Hopfensitz et al. 2010, BoolNet - an R package
 for generation, reconstruction and analysis of Boolean networks
}
\author{
Leshi Chen, leshi, chen@lincolnuni.ac.nz, chenleshi@hotmail.com
}
\keyword{Boolean}
\keyword{Fundamental}
\keyword{Network}
