% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utility_FBN.R
\name{generateFBNInteraction}
\alias{generateFBNInteraction}
\title{generate FBN interaction}
\usage{
generateFBNInteraction(expressionString, genes)
}
\arguments{
\item{expressionString}{An expression string}

\item{genes}{The involved genes contains in the expression string}
}
\value{
A type object of interaction
}
\description{
generate FBN interaction
}
\examples{
coming later
}
