% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_FBN.R
\name{generateSimilarReport}
\alias{generateSimilarReport}
\title{Generate similarity report}
\usage{
generateSimilarReport(similarityreport)
}
\arguments{
\item{similarityreport}{The raw similarity report which was created by the function checkSimilarity}
}
\value{
An organized similarity report
}
\description{
Generate similarity report
}
\examples{
##coming later
}
