% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{generateSimilaryReport}
\alias{generateSimilaryReport}
\title{This method compare two timeseries data and generate similar report}
\usage{
generateSimilaryReport(timeseriesdata1, timeseriesdata2)
}
\arguments{
\item{timeseriesdata1}{the source time series data}

\item{timeseriesdata2}{tge target time series data}
}
\description{
This method compare two timeseries data and generate similar report
}
