% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{genereateBoolNetTimeseries}
\alias{genereateBoolNetTimeseries}
\title{A method to generate BoolNet type of timeseries data}
\usage{
genereateBoolNetTimeseries(
  network,
  initialStates,
  numMeasurements,
  type = c("synchronous", "asynchronous", "probabilistic"),
  geneProbabilities = NULL
)
}
\arguments{
\item{network}{A traditional Boolea type of network that can be used for BoolNet}

\item{initialStates}{A list of initial states}

\item{numMeasurements}{the number of timepoints that need to reconstruct}

\item{type}{the type of the network in traditional Boolean modelling}

\item{geneProbabilities}{optional if type is probabilistic, and then 
it is a need to specify the probabilities for each gene}
}
\description{
A method to generate BoolNet type of timeseries data
}
