% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_FBN.R
\name{getFBMSuccessor}
\alias{getFBMSuccessor}
\title{This method is used to calculate the next state}
\usage{
getFBMSuccessor(
  fbnNetwork,
  previous_states,
  current_step,
  genes,
  type = c("synchronous", "asynchronous"),
  decayIndex = c()
)
}
\arguments{
\item{previous_states}{A vector of current gene state}

\item{genes}{a list of genes which index order must match with the
current state}

\item{type}{A type of Boolean network update schema choosen from synchronous,
asynchronous based. Asynchronous will randomly pick up a gene to process at time.}

\item{decayIndex}{An value indicates the period of time when to degrade 
an activated gene if no activators presented. It is usually one time step}

\item{fbnnetwork}{An object of FBNNetwork}
}
\value{
A list object that contains reconstructed time series and FBN network
}
\description{
This method is used to calculate the next state
}
\examples{
require(BoolNet)
require(FBNNet)
network<-loadNetwork('testthat/others/cellcycle.txt')
trainingseries<-FBNDataReduction(generateTimeSeries(network,2000,43))
cube<-constructFBNCube(network$genes,network$genes,trainingseries,4,1,TRUE)
NETWORK2<-mineFBNNetwork(cube,network$genes)
state<-c('0','1','1','0','1','1','1','1','0','0')
names(state)<-c('CycD','Rb','E2F','CycE','CycA','p27','Cdc20','Cdh1','UbcH10','CycB')
getFBMSuccessor(NETWORK2,state,names(state),'synchronous')
}
