% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getGenePrababilities}
\alias{getGenePrababilities}
\title{The main function to main FBN probabilities from time series data}
\usage{
getGenePrababilities(
  main_parameters_in_ref,
  fixedgenestate,
  target_gene,
  new_conditional_gene,
  temporal,
  targetCounts
)
}
\arguments{
\item{main_parameters_in_ref}{The environment that contains the required data.}

\item{fixedgenestate}{The up stream fixed gene state.}

\item{target_gene}{The target gene.}

\item{new_conditional_gene}{The current stream conditional gene.}

\item{temporal}{The temporal time step.}

\item{targetCounts}{A list of pre-calculated targe genes.}
}
\description{
The main function to main FBN probabilities from time series data
}
