% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getGenePrababilities_advanced}
\alias{getGenePrababilities_advanced}
\title{The main function to main FBN probabilities from time series data}
\usage{
getGenePrababilities_advanced(getGenePrababilities_basic)
}
\arguments{
\item{getGenePrababilities_basic}{The basic calculations.}
}
\description{
The main function to main FBN probabilities from time series data
}
