% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getGenePrababilities_basic}
\alias{getGenePrababilities_basic}
\title{A function to get gene probabilities}
\usage{
getGenePrababilities_basic(
  main_parameters_in_ref,
  fixedgenestate,
  target_gene,
  new_conditional_gene,
  temporal,
  targetCounts
)
}
\arguments{
\item{main_parameters_in_ref}{An environment variable to hold all input data.}

\item{fixedgenestate}{A list of pre-processed gene state.}

\item{target_gene}{The target gene}

\item{new_conditional_gene}{The new conditional gene to be added in.}

\item{temporal}{The temporal time step.}

\item{targetCounts}{The count of the target count.}
}
\description{
A function to get gene probabilities
}
