% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getGenePrababilities_measurements}
\alias{getGenePrababilities_measurements}
\title{Get the main measurements based on the input data}
\usage{
getGenePrababilities_measurements(
  targetGene,
  mainParameters,
  genes,
  matchedgenes,
  temporal = 1L,
  targetCounts = NULL
)
}
\arguments{
\item{targetGene}{The target gene}

\item{mainParameters}{An environment variable holds all input data}

\item{genes}{All conditional genes}

\item{matchedgenes}{processed genes}

\item{temporal}{The temporal time steps}

\item{targetCounts}{}
}
\description{
Get the main measurements based on the input data
}
