% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNBioinformatics.R
\name{getncbiseq}
\alias{getncbiseq}
\title{Retrieving genome sequence data using SeqinR. The code is a sample from the book 'a little book of r for bioinformatic'}
\usage{
getncbiseq(accession)
}
\arguments{
\item{accession}{NBCI accession number}
}
\value{
genome sequence
}
\description{
Retrieving genome sequence data using SeqinR. The code is a sample from the book 'a little book of r for bioinformatic'
}
