% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_FBN.R
\name{getProbabilityFromFunctionInput}
\alias{getProbabilityFromFunctionInput}
\title{An internal method to get the probability of regulatory function from the cube}
\usage{
getProbabilityFromFunctionInput(
  funcType,
  FBNExpression,
  FBNProbability,
  preGeneInputs
)
}
\arguments{
\item{funcType}{type of function}

\item{FBNExpression}{expression of function}

\item{preGeneInputs}{pre input genes' state}

\item{FBNCube}{An object of orchard cube}

\item{targetgene}{The target gene}
}
\value{
A probablity of the target regulatory function
}
\description{
An internal method to get the probability of regulatory function from the cube
}
\examples{
##coming later
}
