% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{getRelatedGeneTimeseries}
\alias{getRelatedGeneTimeseries}
\title{A methiod to reduce the timeseries data based on the gene list}
\usage{
getRelatedGeneTimeseries(timeseries, genelist = c())
}
\arguments{
\item{timeseries}{a list of samples that contains genes on row and time steps on columns}

\item{genelist}{An vector of genes}
}
\description{
A methiod to reduce the timeseries data based on the gene list
}
