% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNBioinformatics.R
\name{getSpecificExpressedGenes}
\alias{getSpecificExpressedGenes}
\title{getSpecificExpressedGenes}
\usage{
getSpecificExpressedGenes(orderSampleTimeSeries, genelist = c())
}
\arguments{
\item{orderSampleTimeSeries}{the original timeseries data that contain continual values}

\item{genelist}{the target genes}
}
\description{
getSpecificExpressedGenes
}
