% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_FBN.R
\name{getStatisticMeasures}
\alias{getStatisticMeasures}
\title{get Statistic Measures}
\usage{
getStatisticMeasures(featurenames, timeseriesdata)
}
\arguments{
\item{featurenames}{a list of feature names, i.e., the gene name}

\item{timeseriesdata}{the original timeseries data that contains continual values}
}
\value{
statstic measures
}
\description{
A method to calculate the statistic measurements based on feature names
}
