% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNBioinformatics.R
\name{identifyDifferentiallyExpressedGenes}
\alias{identifyDifferentiallyExpressedGenes}
\title{A benchmark type function to test a complete process of FBN model
All sub time series must contain the same number of timepoints
Step 3, identify significantly expressed genes, which are strongly related with the samples / study purposes}
\usage{
identifyDifferentiallyExpressedGenes(
  orderSampleTimeSeries,
  cutOffInduction = 1,
  cutOffRepression = 1,
  majority = 7,
  needLog2scale = FALSE,
  probesetGeneNameMappings = NULL,
  nameTab = "RMA"
)
}
\arguments{
\item{orderSampleTimeSeries}{A sorted time series data, which is the output of the method reorderSampleTimeSeries}

\item{cutOffInduction}{a threshold that identify genes as folds. If the cutOffInduction is 2, the differential genes are identified based on 2 folds}

\item{cutOffRepression}{a threshold that identify genes as folds. If the cutOffRepression is 2, the differential genes are identified based on 2 folds}

\item{majority}{A criteria that make a gene as differential}

\item{needLog2scale}{If it is true, then all gene values will be processed using log2}

\item{probesetGeneNameMappings}{gene mapping file}
}
\description{
A benchmark type function to test a complete process of FBN model
All sub time series must contain the same number of timepoints
Step 3, identify significantly expressed genes, which are strongly related with the samples / study purposes
}
