% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_FBN.R
\name{internal_constructCubeWithGroups}
\alias{internal_constructCubeWithGroups}
\title{An internal functions}
\usage{
internal_constructCubeWithGroups(
  index,
  combined_clusters,
  conditional_genes,
  timeseries_data,
  maxK,
  temporal,
  useParallel,
  threshold_confidence,
  threshold_error,
  threshold_support,
  maxFBNRules,
  output_cube
)
}
\arguments{
\item{index}{The current index}

\item{combined_clusters}{The combined clusters}

\item{conditional_genes}{The current conditional genes}

\item{timeseries_data}{The time series data.}

\item{maxK}{The maximum level that can be drilled into}

\item{temporal}{A value that used to be 1 indicates the previous steps the
current one can depend on.}

\item{useParallel}{If it is TRUE, the constructing will run it in parallel,
otherwise in a singl thread.}

\item{threshold_confidence}{A threshold of confidence.}

\item{threshold_error}{A threshold of error.}

\item{threshold_support}{A threshold of support.}

\item{maxFBNRules}{A threshold of maximum rules}

\item{output_cube}{Optional, if it is TRUE, the cube and networks will be
returned, otherwise, only networks are returned.}
}
\description{
An internal functions
}
