% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_FBN.R
\name{isBooleanTypeTimeseriesData}
\alias{isBooleanTypeTimeseriesData}
\title{A simple function to check a value is boolean type
data.}
\usage{
isBooleanTypeTimeseriesData(x)
}
\arguments{
\item{x}{A value that need to be checked.}
}
\value{
TRUE or FALSE
}
\description{
A simple function to check a value is boolean type
data.
}
