% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_FBN.R
\name{isSatisfied}
\alias{isSatisfied}
\title{Is the Gene State Satisfied}
\usage{
isSatisfied(geneState, expression)
}
\arguments{
\item{geneState}{Pre gene input state}

\item{expression}{The expression of the target regulatory function}
}
\value{
TRUE or FALSE
}
\description{
An utility function to verify whether or not the input expression
 and the required gene input are marched
}
\examples{
##coming later
}
