% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utility_FBN.R
\name{isSubExpression}
\alias{isSubExpression}
\title{An internal function to check if an expression is a sub expression}
\usage{
isSubExpression(subExpression)
}
\arguments{
\item{subExpression}{A sub Boolean expression}
}
\description{
An internal function to check if an expression is a sub expression
}
