% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{loadFBNNetwork}
\alias{loadFBNNetwork}
\title{Load FBN network data into a FBN network object}
\usage{
loadFBNNetwork(file, bodySeparator = ",", lowercaseGenes = FALSE)
}
\arguments{
\item{file}{A file that contains FBN data}

\item{bodySeparator}{A specific text separator}

\item{lowercaseGenes}{if all genes are in lower case}
}
\value{
An object of FBN network
}
\description{
Load FBN network data into a FBN network object
}
\examples{
##coming later
}
