% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNBioinformatics.R
\name{mapProbesetNames}
\alias{mapProbesetNames}
\title{This method is used to map probeset names with gene names}
\usage{
mapProbesetNames(probesets, names_mapped = NULL)
}
\arguments{
\item{probesets}{A vector of probeset names}

\item{names_mapped}{the names that have been mapped/processed}
}
\description{
This method is used to map probeset names with gene names
}
