% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBNBioinformatics.R
\name{mapToGeneNameWithhgu133plus2Db}
\alias{mapToGeneNameWithhgu133plus2Db}
\title{Annotation method}
\usage{
mapToGeneNameWithhgu133plus2Db(probesets)
}
\arguments{
\item{probesets}{the target probeset names}
}
\description{
The method maps the probesets with gene names
}
