% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{mergeClusterNetworks}
\alias{mergeClusterNetworks}
\title{A function to merge the networks from clustered FBN cube.}
\usage{
mergeClusterNetworks(clusteredFBNCube, threshold_error, maxFBNRules)
}
\arguments{
\item{clusteredFBNCube}{The clustered FBN cube}

\item{threshold_error}{A threshold of error}

\item{maxFBNRules}{the max FBN rules.}
}
\description{
A function to merge the networks from clustered FBN cube.
}
