% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{mergeInteraction}
\alias{mergeInteraction}
\title{A function to merge the network interactions}
\usage{
mergeInteraction(
  interactions1,
  interactions2,
  genes1,
  genes2,
  mergedgenes = NULL
)
}
\arguments{
\item{interactions1}{the network interaction 1}

\item{interactions2}{the network interaction 1}

\item{genes1}{The genes involved in the interactions1}

\item{genes2}{The genes involved in the interactions2}

\item{mergedgenes}{The merged genes}
}
\description{
A function to merge the network interactions
}
