% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{mergeNetwork}
\alias{mergeNetwork}
\title{A function to merge tow networks.}
\usage{
mergeNetwork(network1, network2)
}
\arguments{
\item{network1}{The first network}

\item{network2}{The second network.}
}
\description{
A function to merge tow networks.
}
