% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_FBN.R
\name{mineFBNNetwork}
\alias{mineFBNNetwork}
\title{Mine FBN Networks from an Orchard cube}
\usage{
mineFBNNetwork(
  fbnGeneCube,
  genes = NULL,
  useParallel = FALSE,
  threshold_confidence = 1,
  threshold_error = 0,
  threshold_support = 1e-05,
  maxFBNRules = 5
)
}
\arguments{
\item{fbnGeneCube}{A pre constructed Orchard cube}

\item{genes}{The target genes in the output}

\item{useParallel}{An option turns on parallel}

\item{threshold_confidence}{A threshod of confidence (between 0 and 1) that used to filter the Fundamental Boolean functions}

\item{threshold_error}{A threshod of error rate (between 0 and 1) that used to filter the Fundamental Boolean functions}

\item{threshold_support}{A threshod of support (between 0 and 1) that used to filter the Fundamental Boolean functions}

\item{maxFBNRules}{The maximum rules per type (Activation and Inhibition) per gene can be mined or filtered, the rest will be discarded}
}
\value{
A object of FBN network
}
\description{
Mine FBN Networks from an Orchard cube
}
\examples{
require(BoolNet)
data("ExampleNetwork")
initialStates <- generateAllCombinationBinary(ExampleNetwork$genes)
trainingseries <- genereateBoolNetTimeseries(ExampleNetwork,
                                           initialStates,
                                           43,
                                           type='synchronous')
cube <- constructFBNCube(ExampleNetwork$genes,
                       ExampleNetwork$genes,
                       trainingseries,
                       4,
                       1,
                       TRUE)
NETWORK <- mineFBNNetwork(cube,ExampleNetwork$genes)
NETWORK
}
\references{
Chen et al.(2018), Front. Physiol., 25 September 2018, 
(\href{https://doi.org/10.3389/fphys.2018.01328}{Front. Physiol.})
}
\author{
Leshi Chen, leshi, chen@lincolnuni.ac.nz
}
\keyword{Boolean}
\keyword{Fundamental}
\keyword{Genetic}
\keyword{Network}
\keyword{Network,}
\keyword{Regulatory}
