% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_FBN.R
\name{mineFBNNetworkStage2}
\alias{mineFBNNetworkStage2}
\title{Internal method}
\usage{
mineFBNNetworkStage2(res, threshold_error = 0, maxFBNRules = 5)
}
\arguments{
\item{res}{The stage 1 result}

\item{threshold_error}{The error threshold}

\item{maxFBNRules}{The maximum FBN rules.}
}
\description{
Internal method
}
