% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mpaste}
\alias{mpaste}
\title{A function concate a vector with a sperator}
\usage{
mpaste(x, sep)
}
\arguments{
\item{x}{The vector}

\item{sep}{Seperator}
}
\value{
A new string
}
\description{
A function concate a vector with a sperator
}
