% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{networkFiltering}
\alias{networkFiltering}
\title{A function to filter networks.}
\usage{
networkFiltering(res)
}
\arguments{
\item{res}{A list of named network interactions}
}
\description{
A function to filter networks.
}
