% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attractor_FBN.R
\name{networkFixUpdate}
\alias{networkFixUpdate}
\title{A method to fix a specific genes in the FBM networks}
\usage{
networkFixUpdate(network, fixIndices, values)
}
\arguments{
\item{network}{A fundamentalBooleanNetwork type of Network}

\item{fixIndices, }{a vector of gene indexes that are intended
to be fixed}

\item{values, }{a vector of values for genes. 0 means fixing
the gene to inhibition 1 means fixing the gene to activation
 and -1 means no fixing}
}
\description{
A method to fix a specific genes in the FBM networks
}
