% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_FBN.R
\name{output_annotated_genes}
\alias{output_annotated_genes}
\title{A method to convert a vector of gene names to annotated gene details}
\usage{
output_annotated_genes(genes, path, filename)
}
\arguments{
\item{genes}{A vector of genes}

\item{path}{The file path}

\item{filename}{The name of the output file such as xx.csv}
}
\description{
A method to convert a vector of gene names to annotated gene details
}
