% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_FBN.R
\name{output_timesereis_based_on_genes}
\alias{output_timesereis_based_on_genes}
\title{A method to reorder a matrix by a vector}
\usage{
output_timesereis_based_on_genes(matrix, genes)
}
\arguments{
\item{matrix}{A matrix}

\item{genes}{The target genes in the row names of the matrix}
}
\description{
A method to reorder a matrix by a vector
}
