% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_FBN.R
\name{plotNetwork}
\alias{plotNetwork}
\title{A function to plot the Fundamental boolean networks}
\usage{
plotNetwork(
  FBNNetwork,
  target_genes = c(),
  type = c("static", "staticSlice", "dynamic", "forward_1a", "forward_2a", "forward_3a",
    "forward_4a", "backward_1a", "backward_2a", "forward_1b", "forward_2b", "forward_3b",
    "forward_4b", "backward_1b", "backward_2b"),
  expand_level = 2,
  output_network = FALSE,
  timeseries_matrix = NULL,
  start_time_point = 1,
  end_time_point = 1,
  target_time_point = 1
)
}
\arguments{
\item{FBNNetwork}{The FBN networks}

\item{target_genes}{A list of target genes}

\item{type}{The visualization type}

\item{expand_level}{The level of expandation}

\item{output_network}{Optional, if TRUE, then output the FBN Network}

\item{timeseries_matrix}{The timeseries matrix that used to plot dynamic 
networks.}

\item{start_time_point}{The start time point}

\item{end_time_point}{The end time point}

\item{target_time_point}{The target time point}
}
\description{
A function to plot the Fundamental boolean networks
}
