% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FBMAttractors.R
\name{print.FBMAttractors}
\alias{print.FBMAttractors}
\title{A S3 method to print the FBM attractors}
\usage{
\method{print}{FBMAttractors}(x, ...)
}
\arguments{
\item{x}{The attactor details}

\item{...}{other arguments.}
}
\description{
A S3 method to print the FBM attractors
}
