% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FundamentalBooleanNetwork.R
\name{print.FundamentalBooleanNetwork}
\alias{print.FundamentalBooleanNetwork}
\title{A S3 method to print the FBM network}
\usage{
\method{print}{FundamentalBooleanNetwork}(x, ...)
}
\arguments{
\item{x}{The network details}

\item{...}{other arguments.}
}
\description{
A S3 method to print the FBM network
}
