% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{process_cube_algorithm}
\alias{process_cube_algorithm}
\title{Get the main measurements based on the input data}
\usage{
process_cube_algorithm(
  target_gene,
  conditional_genes,
  maxK,
  temporal,
  mainParameters
)
}
\arguments{
\item{target_gene}{The target gene}

\item{conditional_genes}{conditional genes}

\item{maxK}{The maximum under ground levels}

\item{temporal}{The temporal time steps}

\item{mainParameters}{An environment variable holds all input data}
}
\description{
Get the main measurements based on the input data
}
