% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{randomGenerateBinary}
\alias{randomGenerateBinary}
\title{A method to generate binary data randomly}
\usage{
randomGenerateBinary(genelist = c(), maxState = 0)
}
\arguments{
\item{genelist}{An vector of genes}

\item{maxState}{that should be less tha 2^length(genelist)}
}
\description{
A method to generate binary data randomly
}
\examples{
individualgenes<-c('a','b','c')
testdata2 <- randomGenerateBinary(individualgenes,maxState = 4)
testdata2
}
