% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_FBN.R
\name{randomSelection}
\alias{randomSelection}
\title{A random selection function denoted as P[]}
\usage{
randomSelection(probability)
}
\arguments{
\item{probability}{A probability of an event}
}
\value{
TRUE or FALSE indicating whether or not the event has been selected
}
\description{
A random selection function denoted as P[]
}
\examples{
coming later
}
