% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{rbind.all.columns}
\alias{rbind.all.columns}
\title{A method to bind two matrix by row}
\usage{
\method{rbind}{all.columns}(x, y)
}
\arguments{
\item{x}{matrix one}

\item{y}{matrix two}
}
\description{
A method to bind two matrix by row
}
