% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{rbind_all_columns}
\alias{rbind_all_columns}
\title{A method to bind two matrix by row}
\usage{
rbind_all_columns(x, y)
}
\arguments{
\item{x}{matrix one}

\item{y}{matrix two}
}
\description{
A method to bind two matrix by row
}
