% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_FBN.R
\name{reconstructTimeseries}
\alias{reconstructTimeseries}
\title{This method is used to reconstruct time series data}
\usage{
reconstructTimeseries(
  fbnnetwork,
  initialStates,
  type = c("synchronous", "asynchronous"),
  maxTimepoints = 100,
  useParallel = FALSE
)
}
\arguments{
\item{fbnnetwork}{An object of FBNNetwork}

\item{initialStates}{A list of initial states}

\item{type}{Specify the type of the Fundamental Boolean model (synchronous or asynchronous)}

\item{maxTimepoints}{The max time points that are going to be constructed}
}
\value{
A list object that contains reconstructed time series and FBN network
}
\description{
This method is used to reconstruct time series data
}
\examples{
mat1<-matrix(c('1','0','0','1','0','0','0','1','1'),3,3, dimnames=list(c('gene1','gene2','gene3'),c('1','2','3')))
mat2<-matrix(c('1','1','0','1','0','1','1','1','0'),3,3, dimnames=list(c('gene1','gene2','gene3'),c('1','2','3')))
listtestOriginal<-list(mat1,mat2)
cube<-constructFBNCube(c('gene1','gene2'),c('gene1','gene2','gene3'),listtestOriginal,4,1,FALSE)
network<-mineFBNNetwork(cube,c('gene1','gene2'))
mat1<-matrix(c('1','1','0'),3,1, dimnames=list(c('gene1','gene2','gene3'),c('1')))
mat2<-matrix(c('1','1','1'),3,1, dimnames=list(c('gene1','gene2','gene3'),c('1')))
listtestInitial<-list(mat1,mat2)
result<-reconstructTimeseries(network,listtestInitial,1,'synchronous')
}
