% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utility_FBN.R
\name{regenerateInteractions}
\alias{regenerateInteractions}
\title{An internal function to reconstrunct FBN functions}
\usage{
regenerateInteractions(
  name,
  expressionstring,
  genes,
  error,
  type,
  probability = NA,
  support = NA,
  timestep = 1
)
}
\arguments{
\item{name}{The name of the interaction}

\item{expressionstring}{The string presentation of an expression.}

\item{genes}{The involve genes}

\item{error}{The error value}

\item{type}{The type of the interaction}

\item{probability}{The probability of the interaction}

\item{support}{The support threshold.}

\item{timestep}{The time step of the interaction to take effectiveness on.}
}
\description{
An internal function to reconstrunct FBN functions
}
