% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_FBN.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Internal method}
\usage{
removeDuplicates(factors)
}
\arguments{
\item{factors}{The value that needs to be duplicated.}
}
\description{
Internal method
}
