% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_FBN.R
\name{reorderSampleTimeSeries}
\alias{reorderSampleTimeSeries}
\title{This method is used to sort time series order based on columns}
\usage{
reorderSampleTimeSeries(
  convertedTimeSeries,
  func = function(x) x[length(x)],
  splitor = as.character("-")
)
}
\arguments{
\item{convertedTimeSeries}{An output of the method 
convertIntoSampleTimeSeries that converts normalized timeseries data
into a list of samples}

\item{func}{A function that specified how to split the column names}

\item{splitor}{Separator.}
}
\description{
This method is used to sort time series order based on columns
}
