% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{rep.row}
\alias{rep.row}
\title{A method to duplicate a row}
\usage{
\method{rep}{row}(x, n)
}
\arguments{
\item{x}{the row value that need to be repeated}

\item{n}{the number of repeats}
}
\description{
A method to duplicate a row
}
