% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{rep_col}
\alias{rep_col}
\title{A method to duplicate a column}
\usage{
rep_col(x, n)
}
\arguments{
\item{x}{the column value that need to be repeated}

\item{n}{the number of repeats}
}
\description{
A method to duplicate a column
}
