% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_FBN.R
\name{similarityBetweenMatrix}
\alias{similarityBetweenMatrix}
\title{This function is used to compare the similarity of two matrixes}
\usage{
similarityBetweenMatrix(timeseries1, timeseries2, index)
}
\arguments{
\item{timeseries1}{A matrix}

\item{timeseries2}{The second matrix to compare with}

\item{index}{A label to distinguish the result.}
}
\value{
A vector to show the similarity information about the two matrixes.
}
\description{
This function is used to compare the similarity of two matrixes
}
