% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{splitExpression}
\alias{splitExpression}
\title{A function to split an expression into a vector of input}
\usage{
splitExpression(expression, outputType, lowerCase)
}
\arguments{
\item{expression}{The expression fo a FBN network connection}

\item{outputType}{The type of output.}

\item{lowerCase}{Optional, if TRUE convert them to lower case.}
}
\description{
A function to split an expression into a vector of input
}
