% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_FBN.R
\name{StaticNetwork}
\alias{StaticNetwork}
\title{Display a static network in a slice}
\usage{
StaticNetwork(networkobject, david_gene_list = NULL)
}
\arguments{
\item{networkobject}{A FBN network object}

\item{david_gene_list}{the list object of DAVID annotation gene list}
}
\value{
No return
}
\description{
Display a static network in a slice
}
\examples{
##coming later
}
