% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_FBN.R
\name{transitionStates}
\alias{transitionStates}
\title{A function do the state transition}
\usage{
transitionStates(
  initialState,
  fbnNetwork,
  genes,
  type = c("synchronous", "asynchronous"),
  maxTimepoints
)
}
\arguments{
\item{initialState}{The initial state}

\item{fbnNetwork}{The FBN network}

\item{genes}{The involved genes}

\item{type}{The type of Boolean transition}

\item{maxTimepoints}{The maximum timepoints need to be 
generated.}
}
\description{
A function do the state transition
}
