% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeastTimeSeries.R
\docType{data}
\name{yeastTimeSeries}
\alias{yeastTimeSeries}
\title{Yeast Timeseries Data}
\format{
A simple yeast timeseries
}
\usage{
data(yeastTimeSeries)
}
\description{
Data generated for testing purpose.
}
\examples{
data(yeastTimeSeries)
yeastTimeSeries
}
\references{
Mussel, Hopfensitz et al. 2010, BoolNet - an R package
 for generation, reconstruction and analysis of Boolean networks
}
\keyword{Yeast}
