% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utility_FBN.R
\name{convertMinedResultToFBNNetwork}
\alias{convertMinedResultToFBNNetwork}
\title{An internal function to convert the mined result to FBN network}
\usage{
convertMinedResultToFBNNetwork(minerresult, genes)
}
\arguments{
\item{minerresult}{The result of mining.}

\item{genes}{The genes involved in the mining.}
}
\description{
An internal function to convert the mined result to FBN network
}
