% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAVID_gene_list.R
\docType{data}
\name{DAVID_gene_list}
\alias{DAVID_gene_list}
\title{DAVID Gene Annotation data}
\format{
A table of gene list mapping probeset with gene names extracted from DAVID tools.
}
\source{
\href{https://david.ncifcrf.gov/}{DAVID Bioinformatic Database}
}
\usage{
data(DAVID_gene_list)
}
\description{
Data extracted from DAVID tools for mapping probeset ids with gene names.
}
\examples{
data(DAVID_gene_list)
DAVID_gene_list
}
\references{
Nature Protocols 2009; 4(1):44 & Nucleic Acids Res. 2009;37(1):1 
(\href{https://david.ncifcrf.gov/}{DAVID Bioinformatic Database})
}
\keyword{Annotation}
\keyword{Gene}
