% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_FBN.R
\name{FBNNetwork.Graph}
\alias{FBNNetwork.Graph}
\title{Executing a list of processes in parallel with decreasing list}
\usage{
FBNNetwork.Graph(
  fbnNetwork,
  type = "static",
  timeseriesMatrix = NULL,
  fromTimePoint = 1,
  toTimePoint = 5,
  networkobject = NULL
)
}
\arguments{
\item{fbnNetwork}{A FBN Network object}

\item{type}{A type of object and its value is type=static | dynamic | staticSlice}

\item{timeseriesMatrix}{A sample timeseries matrix}

\item{fromTimePoint}{The time point at the beginning}

\item{toTimePoint}{The time point at the end}

\item{networkobject}{An object created by the funciton
\code{ConvertToNetworkGraphicObject}.}
}
\value{
No return
}
\description{
Executing a list of processes in parallel with decreasing list
}
\examples{
##coming later
}
