% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{filterNetworkConnectionsByGenes}
\alias{filterNetworkConnectionsByGenes}
\title{Find Backward Network By Genes}
\usage{
filterNetworkConnectionsByGenes(
  networks,
  genelist = c(),
  exclusive = TRUE,
  expand = FALSE
)
}
\arguments{
\item{networks}{The Fundamental Boolean Network internally}

\item{genelist}{A list of genes}

\item{exclusive}{Optional}

\item{expand}{Optional}
}
\description{
Find Backward Network By Genes
}
