% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_application_FBN.R
\name{findBackwardRelatedNetworkByGenes}
\alias{findBackwardRelatedNetworkByGenes}
\title{Find Backward Network By Genes}
\usage{
findBackwardRelatedNetworkByGenes(
  networks,
  target_gene_list = c(),
  regulationType = NULL,
  maxDeep = 1,
  next_level_mix_type = FALSE
)
}
\arguments{
\item{networks}{The Fundamental Boolean Network}

\item{target_gene_list}{The target genes}

\item{regulationType}{The type of regulation either in 1 (Up regulation) or 0 (down regulation)}

\item{maxDeep}{The maximum layous that need to be drilled down.}

\item{next_level_mix_type}{If TRUE, the next level will contain all type of connections.}
}
\description{
Find Backward Network By Genes
}
