% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utility_FBN.R
\name{generateAllCombinationBinary}
\alias{generateAllCombinationBinary}
\title{generate all combaination binary data based on an vector of genes}
\usage{
generateAllCombinationBinary(genelist = c(), begin = 1, last = 0)
}
\arguments{
\item{genelist}{An vector of genes}

\item{begin}{return begin index}

\item{last}{return last index}
}
\description{
generate all combaination binary data based on an vector of genes
}
