% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_FBN.R
\name{getFBMSuccessor}
\alias{getFBMSuccessor}
\title{This method is used to calculate the next state}
\usage{
getFBMSuccessor(
  fbnNetwork,
  previous_states,
  current_step,
  genes,
  type = c("synchronous", "asynchronous"),
  decayIndex = c()
)
}
\arguments{
\item{fbnNetwork}{An object of FBNNetwork}

\item{previous_states}{A vector of current gene state}

\item{current_step}{The index of the current step}

\item{genes}{a list of genes which index order must match with the
current state}

\item{type}{A type of Boolean network update schema choosen from synchronous,
asynchronous based. Asynchronous will randomly pick up a gene to process at time.}

\item{decayIndex}{An value indicates the period of time when to degrade 
an activated gene if no activators presented. It is usually one time step}
}
\value{
A list object that contains reconstructed time series and FBN network
}
\description{
This method is used to calculate the next state
}
\examples{
require(BoolNet)
data(ExampleNetwork)
trainingseries<-FBNDataReduction(generateTimeSeries(ExampleNetwork,2000,43))
cube<-constructFBNCube(target_genes = ExampleNetwork$genes,
                       conditional_genes = ExampleNetwork$genes,
                       timeseriesCube = trainingseries,
                       maxK = 4,
                       temporal = 1,
                       useParallel = FALSE)
NETWORK2<-mineFBNNetwork(cube,ExampleNetwork$genes)
state<-c('0','1','1','0','1')
names(state)<-c('Gene1','Gene2','Gene3','Gene4','Gene5')
getFBMSuccessor(NETWORK2, 
                previous_states= state,
                current_step = 2,
                genes = names(state),
                type = 'synchronous')
}
