% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utility_FBN.R
\name{isAppliedDeMorganLaw}
\alias{isAppliedDeMorganLaw}
\title{An internal function to check a splited Boolean expression
is applied DeMorganLaw}
\usage{
isAppliedDeMorganLaw(splitedExpression)
}
\arguments{
\item{splitedExpression}{A splited Boolean expression}
}
\description{
An internal function to check a splited Boolean expression
is applied DeMorganLaw
}
