% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_FBN.R
\name{mineFBNNetworkWithCores}
\alias{mineFBNNetworkWithCores}
\title{An internal function}
\usage{
mineFBNNetworkWithCores(
  searchFBNNetworkCore,
  genes = NULL,
  threshold_error,
  maxFBNRules
)
}
\arguments{
\item{searchFBNNetworkCore}{A result of \code{searchFBNNetworkCore}}

\item{genes}{Genes that involved.}

\item{threshold_error}{A threshod of error rate (between 0 and 1) 
that used to filter the Fundamental Boolean functions}

\item{maxFBNRules}{The maximum rules per type (Activation and Inhibition) 
per gene can be mined or filtered, the rest will be discarded}
}
\description{
An internal function
}
