% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_FBN.R
\name{reconstructTimeseries}
\alias{reconstructTimeseries}
\title{This method is used to reconstruct time series data}
\usage{
reconstructTimeseries(
  fbnnetwork,
  initialStates,
  type = c("synchronous", "asynchronous"),
  maxTimepoints = 100,
  useParallel = FALSE
)
}
\arguments{
\item{fbnnetwork}{An object of FBNNetwork}

\item{initialStates}{A list of initial states}

\item{type}{Specify the type of the Fundamental Boolean model (synchronous or asynchronous)}

\item{maxTimepoints}{The max time points that are going to be constructed}

\item{useParallel}{Optional, if TRUe then use parallelisation}
}
\value{
A list object that contains reconstructed time series and FBN network
}
\description{
This method is used to reconstruct time series data
}
\examples{
require(BoolNet)
data('ExampleNetwork')
initialStates <- generateAllCombinationBinary(ExampleNetwork$genes)
trainingseries <- genereateBoolNetTimeseries(ExampleNetwork,
                                           initialStates,
                                           43,
                                           type='synchronous')
cube <- constructFBNCube(target_genes = ExampleNetwork$genes,
                       conditional_genes = ExampleNetwork$genes,
                       timeseriesCube = trainingseries,
                       maxK = 4,
                       temporal = 1,
                       useParallel = FALSE)
network <- mineFBNNetwork(cube)
result <- reconstructTimeseries(fbnnetwork = network,
                              initialStates = initialStates,
                              type = 'synchronous',
                              maxTimepoints = 43)
result
}
