% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utility_FBN.R
\name{removeFistLevelBracket}
\alias{removeFistLevelBracket}
\title{An internal function to remove the first brasket}
\usage{
removeFistLevelBracket(splitedExpression)
}
\arguments{
\item{splitedExpression}{A splited Boolean expression}
}
\description{
An internal function to remove the first brasket
}
