% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attractor_FBN.R
\name{searchForAttractors}
\alias{searchForAttractors}
\title{A function to find all possible FBM (Fundamental Boolean models) attractors}
\usage{
searchForAttractors(
  fbnNetwork,
  startStates = list(),
  genes,
  type = c("synchronous", "asynchronous"),
  genesOn = c(),
  genesOff = c(),
  maxSearch = 1000
)
}
\arguments{
\item{fbnNetwork}{An object of FBNNetwork}

\item{startStates}{A list of initial states, the row names of each state must
be matched with the genes}

\item{genes}{a list of genes which index order must match with the current
state}

\item{type}{A type of Boolean network update schema chosen from synchronous,
asynchronous and time step based}

\item{genesOn}{It is a vector of genes that are marked as On}

\item{genesOff}{It is a vector of genes that are marked as Off}

\item{maxSearch}{The maximum timesteps that the system will try to search.}
}
\value{
Attractor objects
}
\description{
A function to find all possible FBM (Fundamental Boolean models) attractors
}
\examples{
data('ExampleNetwork')
initialStates <- generateAllCombinationBinary(ExampleNetwork$genes)
trainingseries <- genereateBoolNetTimeseries(ExampleNetwork,
                                           initialStates,
                                           43,
                                           type='synchronous')
cube<-constructFBNCube(target_genes = ExampleNetwork$genes,
                       conditional_genes = ExampleNetwork$genes,
                       timeseriesCube = trainingseries,
                       maxK = 4,
                       temporal = 1,
                       useParallel = FALSE)
NETWORK2 <- mineFBNNetwork(cube,ExampleNetwork$genes)

## find attractor with type = synchronous
attractor <- searchForAttractors(NETWORK2,
                               initialStates,
                               ExampleNetwork$genes)
print(attractor)
FBNNetwork.Graph.DrawAttractor(NETWORK2,attractor,3)

## find attractor with type = asynchronous
attractor <- searchForAttractors(NETWORK2,
                               initialStates,
                               type = 'asynchronous',
                               ExampleNetwork$genes)
print(attractor)
FBNNetwork.Graph.DrawAttractor(NETWORK2,attractor,3)

}
\references{
Chen et al.(2018), Front. Physiol., 25 September 2018, 
(\href{https://doi.org/10.3389/fphys.2018.01328}{Front. Physiol.})

Mussel, Hopfensitz et al. 2010, 
 BoolNet - an R package for generation, reconstruction and analysis of
  Boolean networks
}
\author{
Leshi Chen, leshi, chen@lincolnuni.ac.nz, chenleshi@hotmail.com
}
\keyword{Boolean}
\keyword{Fundamental}
\keyword{Network}
